backend/pyproject.toml
[project]
name = "inventory-backend"
version = "0.1.0"
description = "Inventory auth and sync service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.22",
    "pydantic-settings>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1,<5",
    "pyjwt>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
    "pytest-asyncio>=0.23.7",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
